third_party/blink/web_tests/svg/dynamic-updates/SVGPatternElement-dom-height-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' attribute of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.getAttribute('height') is "100"
PASS patternElement.getAttribute('height') is "50"
PASS successfullyParsed is true

TEST COMPLETE
